pyproject.toml
[project]
name = "dont-block-ping-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "pytest",
    "httpx",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "ruff>=0.9.4",
]
